requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-JWT==0.3.2
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==1.4.2
pymongo==4.11.1
python-dotenv==1.0.1
Werkzeug==3.1.3
